requirements.txt
azure-identity
langchain
langchain-community
azure-search-documents
openai
python-dotenv
fastapi
uvicorn[standard]
tiktoken
sentence_transformers
pandas 
fastapi 
faiss-cpu 
numpy 
qdrant-client
transformers 
pillow 
torch==2.4.0 
torchvision==0.19.0 
torchaudio==2.4.0
librosa
torchaudio
opencv-python
pydub
moviepy
scikit-learn

